requirements.txt
Django==4.1.4
djangorestframework==3.14.0
pytest==7.2.0
pytest-django==4.5.2
black==22.12.0